moved operator costs
